pyproject.toml
[tool.poetry]
name = "seed-backend-python"
version = "0.1.0"
description = ""
authors = ["Karen"]
readme = "README.md"
packages = [
    { include = "*", from = "app" },
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.112.0"}
motor = "^3.5.1"
pydantic = {extras = ["email"], version = "^2.8.2"}
pyjwt = "^2.9.0"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
